// packages/amplify-graphql-api-construct-tests/src/types.ts
/**
 * Represents stack outputs generated by Amplify Auth Construct.
 */
export type AuthConstructStackOutputs = {
  authRegion: string;
  userPoolId: string;
  webClientId: string;
  identityPoolId: string;
};

/**
 * Represents stack outputs generated by AmplifyGraphqlApi Construct and UserPool Construct.
 */
export type UserPoolAuthConstructStackOutputs = {
  apiEndpoint: string;
  apiKey: string;
  region: string;
  userPoolId?: string;
  webClientId?: string;
};